+++ as-api/pkg/api/apiserver/v1/order.go
@@ -9,20 +9,20 @@ type Order struct {
 	// Standard object's metadata.
 	metav1.ObjectMeta
 
-	ProjectID     uint64 `json:"project_id" gorm:"column:project_id;" validate:"required"`
-	VehicleID     uint64 `json:"vehicle_id" gorm:"column:vehicle_id;" validate:"required"`
-	DriverID      uint64 `json:"driver_id" gorm:"column:driver_id;" validate:"required"`
-	Weight        int    `json:"weight" gorm:"weight"`
-	Mileage       int    `json:"mileage" gorm:"column:mileage;"`
-	Freight       int    `json:"freight" gorm:"column:freight;"`
-	Payroll       int    `json:"payroll" gorm:"column:payroll;"`
-	Status        int    `json:"status" gorm:"column:status;"`
-	LoadAt        string `json:"load_at" gorm:"column:load_at;"`
-	UnLoadAt      string `json:"unload_at" gorm:"column:unload_at;"`
-	Comment       string `json:"comment" gorm:"column:comment;" validate:"min=1,max=255"`
-	VehicleNumber string `json:"vehicle_number" gorm:"->"`
-	Project       string `json:"project" gorm:"->"`
-	Driver        string `json:"driver" gorm:"->"`
+	ProjectID     uint64  `json:"project_id" gorm:"column:project_id;" validate:"required"`
+	VehicleID     uint64  `json:"vehicle_id" gorm:"column:vehicle_id;" validate:"required"`
+	DriverID      uint64  `json:"driver_id" gorm:"column:driver_id;" validate:"required"`
+	Weight        float32 `json:"weight" gorm:"weight"`
+	Mileage       float32 `json:"mileage" gorm:"column:mileage;"`
+	Freight       int     `json:"freight" gorm:"column:freight;"`
+	Payroll       int     `json:"payroll" gorm:"column:payroll;"`
+	Status        int     `json:"status" gorm:"column:status;"`
+	LoadAt        string  `json:"load_at" gorm:"column:load_at;"`
+	UnLoadAt      string  `json:"unload_at" gorm:"column:unload_at;"`
+	Comment       string  `json:"comment" gorm:"column:comment;" validate:"min=1,max=255"`
+	VehicleNumber string  `json:"vehicle_number" gorm:"->"`
+	Project       string  `json:"project" gorm:"->"`
+	Driver        string  `json:"driver" gorm:"->"`
 }
 
 // OrderList is the whole list of all orders which have been stored in storage.
